feat: Add message when task is empty
This commit adds a new localized string "taskIsEmpty" to notify users when their task list is empty.

The `TodoListCubit` now has a `taskIsEmpty` method to check if both uncompleted and completed todos are empty.

The `TodoListPage` uses this new message to display it in the center of the screen when no tasks are present.

Additionally, after deleting a task in `TodoListView`, `fetchTodos` is called to update the list.